Buildbot support: Make test failure messages available to scripts
We add a new low-level function, (command-line-error msg), which displays
error messages on std::cerr (or the closest native equivalent).  We call
this function from the test driver to display test case failures.

We also add a 'rake halyard:test' command to the standard list of rake
tasks, and modify the test driver to immediately exit the application after
errors occur.