Save number of files loaded to an ignroable "cached configuration" file
We have been saving the number of files loaded to our
application.halyard, in order to support our progress bar.  The
problem is that this file changes all the time, which is annoying
because it can cause accidental merge conflicts when no real conflicts
exist.

To fix this, we introduce the notion of a "cached configuration" file.
This is found in config/cached.conf, and can be ignored in version
control, but released along with the program to support the progress
bar.

The cached configuration file uses wxFileConfig to write the file out,
so it is in conf file format.  HalyardApp creates an appropriate object
which is registered with TInterpreterManager, which is then used by
TInterpreter.  We also remove all knowledge of the current Document from
TInterpreterManager, since it used it only for the source file count.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?10115