Save separate file counts for lazy and non-lazy loads
We have been disabling saving of file counts when lazy loading is enabled,
because you'll get lower file counts that won't be correct when you're doing
a full load.  This patch adds a new key, LazySourceFileCount, that is used
for saving the file count when lazy loading is enabled, so if we do enable
lazy loading by default, we can still have working progress bars.