Lazy loading: Don't lose track of current card after errors
The scenario:

 1) The user is on card /foo/bar.
 2) The user edits foo.ss, introduces an error, and reloads.
 3) The user sees an error message.
 4) The user fixes foo.ss and reloads.

At the end of this sequence, the user should still be on /foo/bar.  But the
existing version of the engine jumps to /start immediately after step (3),
forcing the user to dig through the program in search of their card after
step (4).

Our fix, unfortunately, is rather drastic.  When a lazy load fails, we
simply shut down the interpreter, and wait for the next reload.  This
preserves the current card name.

This fix also affects the related scenario, where the program is loaded,
and the user browses into a broken group.  Before, this left the program in
an inconsistent state.  Now we just shut down the interpreter.  We'll have
to see how well this works in practice.

In this patch, we add a new script-load-error function which displays an
error message and shuts down the interpreter.  We also turn the
caught_error variable in TInterpreterManager::Run into a member variable,
so that other functions can set mLoadScriptFailed and mCaughtError failed
to true to signal a load error.

We also move a check of mDone from TInterpreterManager::LoadAndRunScript
(where it is no longer useful, because the RunInitialCommands function
hasn't been called yet) into %run-main-kernel-loop, where the check
actually does something.  We also change the initial values of
*%kernel-exit-to-top-func* and *%kernel-exit-interpreter-func* to something
a bit more descriptive.

Please keep a careful eye on script reloading and error handling: This is a
complicated patch, and touches a lot of very low-level internals.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?10554