Editor: Load syntax highlighting information even when first load fails
We had been checking for whether we had an interpreter before
initializing the ScriptEditorDB that read syntax highlighting
information from definitions.sqlite3.  This was more than we actually
require; we can still read the database when we don't have an
interpreter.  This patch fixes the check to ensure only that the
script has begun (which means we know which application.halyard we're
looking at).

This patch also fixes a long-standing bug in which we would dump our
identifier information every time we initialized the ScriptEditorDB
because of a bug in the schema initialization code.

Reviewed-by: Eric Kidd <eric.m.kidd@dartmouth.edu>