Buildbot support: Display Caution messages on console
When in COMMAND_LINE mode, display Caution messages on the console.  We
want to show these in AUHTORING mode, too, but let's hold off until after
we get a build of 0.5.15.x.

Note that we had to modify a Caution call in FileSystem.cpp to reduce
annoying chatter on the Mac.  There's a bug filed to investigate this
problem further, but it doesn't look too urgent.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?10545
Bug: http://iml.dartmouth.edu/fogbugz/default.php?10630