Decrease LOG_APAGE_SIZE to 16 after build failure
A LOG_APAGE_SIZE of 17 seems to cause the engine to crash every time on
my machine.  Thus, in order to get a build, and have a stable engine,
it looks like we're going to have to reduce this back down to 16.

Fix-for: df8205a778ca564a3200785ed459f561f525df21