Buildbot support: Add git:force_clean Rake task
Since fresh checkouts are expensive, and we want quick results from
buildbot, we want to be able to run 'git update' and build the resulting
tree.

But we also want clean, reproducible builds.  This new Rake task is a
compromise: It uses git operations to sanitize a tree completely.  It will
probably fail, however, if filenames change case or submodules move around.