Quake 2: Add support for teleporting when developing levels
This teleport support is for debugging purposes only--it includes visual
distortions and sound effects, and we don't understand all the details of
how it interacts with the underlying Quake 2 code.

We add two new console commands to the Quake 2 game AI:
  pos: Print x, y, z and yaw
  teleport: Zap player to specified x, y, z and yaw

We also add a new console command to Halyard:
  goto: Teleport to a named location

You can define named teleport locations as follows:

  (define-quake2-teleport-location bridge (-55 245 22 90)
    (quake2-command "say Teleporting to bridge."))

On the Quake console, type:

  goto bridge

See the documentation in halyard/quake2/teleport.ss for further details.

Paired-with: Brian Campbell <brian.p.campbell@dartmouth.edu>
Bug: http://iml.dartmouth.edu/fogbugz/default.php?11221