Rename NON-FATAL-ERROR -> REPORT-ERROR
This function was poorly named, because it could actually be fatal (when
not in AUTHORING mode).  It corresponds to gLog.Error in C++, but we can't
call it ERROR, because that name is taken in Scheme.  So we name it after
its most common use: Reporting errors to the user.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11371