buildbot: Force WAIT to return immediately in COMMAND_LINE mode
When running under buildbot, we're going to want to jump to different cards
various cards in the program, run them, and see if we encounter errors.
But there's no point in WAITing through long media elements.  So in
COMMAND_LINE mode, we act as though the user pressed Esc at the start of
each media element.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11396