buildbot: Make it easier to call JUMP from the command-line
Ultimately, we want to make it possible to jump to every card under the
control of the buildbot.  But right now, it's unnecessarily hard to call
JUMP from the command-line, either directly or indirectly.  For example,
the code for the unit test driver has relied on a very cumbersome
workaround to support the following:

  Halyard -c '(command-line-test-driver)'

This required patching into the node system in a rather ugly way.

This patch greatly simplifies how the -c and -e command-line flags work.
We move the "initial command" code inside %main-kernel-loop, where it has
access to ordinary JUMP and EXIT-SCRIPT commands, and restructure
(command-line-test-driver) and TInterpreter.cpp accordingly.

We also add the ability to invoke the GUI test suite from the command-line
on the Mac.  And to help reduce the risk of regressions, I've tested the
"reload script" command with both regular and lazy loading, and made sure
that we jump back to the correct card after an error.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11253
Reviewed-by: Brian Campbell <brian.p.campbell@dartmouth.edu>