mzscheme: Display error dialog when VirtualAlloc fails
The mzscheme garbage collector uses VirtualAlloc to allocate large amounts
of heap space.  But on various versions of Windows, VirtualAlloc may fail
once a certain number and/or total size of pages have been allocated.

PLT v372 doesn't actually check for VirtualAlloc failures, which results in
segfaults in various parts of the garbage collector.  We want to insert a
an actual check here, with a distinctive error message.  This will allow us
to distinguish VirtualAlloc problems (which we have little control over)
from any other kind of crash which might occur in the garbage collector.

Reviewed-by: Robinson Tryon <robinson.p.tryon@dartmouth.edu>
Bug: http://iml.dartmouth.edu/fogbugz/default.php?11009