Jumping to each card: Don't end playback of looping audio
When running in COMMAND_LINE mode, we attempted to immediately mark all
audio playback as completed.  However, this made no sense when called on
looping audio, which should obviously never be marked as completed, even if
we're fast-forwarding through the program.

The old code triggered an assertion failure here:

  ASSERT(mEventDispatcher.get())

This occured because we were attempting to send a MEDIA-FINISHED event to a
GeigerSynthElement, which has no associated event handlers.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11607
Fix-for: cb030b74dea7c359a2514ac56fb90c8b3cf7a347