Bug fix: Re-enable FancyCrashReporter on Windows
When porting the GUI to the Macintosh, I accidentally mispelled a
preprocessor #define and disabled the FancyCrashReporter dialogs on
Windows.  Since the server-side of the crash reporter has been down for a
while now, we never took the time to find and fix this mistake.

Fix-for: c575488a599f45bb9a7118859efeb974e8d74e20
Bug: http://iml.dartmouth.edu/fogbugz/default.php?3112