Clear gStateDB when script is reloaded
When working with the state-db in Scheme, we often rename or change
state-db keys.  But when doing so, we occasionally forget to update the
associated initialization code.  Unfortunately, reloading the script has
not historically cleared gStateDB, so stale state-db entries would prevent
us from seeing any errors.  And this meant we were likely to check in
broken code.

This patch clears TStateDB after making the usual
NotifyReloadScriptStarting calls, and re-initializes it with just the
standard /system/clock/* keys.  This should be a reasonable thing to do,
because--after all--reloading the script is supposed to restore us to a
pristine condition (except for define/p values), and jump to whatever card
we were on last.  So we _do_ run all the usual .SETUP and .INITIALIZE
methods, which is where people are supposed to be initializing the
state-db.