Trap unexpected JUMPs while running unit tests
When unit-testing code, an unexpected call to (jump ...) or (jump-next)
will simply end the unit test.  This is suboptimal, because it's possible
to introduce accidental bugs in button-related code, and not notice that
some code is calling (jump-next) unexpectedly.

We add some low-level code which intercepts jumps in unit tests and turns
them into test-case failures.  Note that we patch this directly into
%test-case% (and not a subclass), because we want it to affect all unit
tests.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?11909