0.5.21 - 28 Oct 2008
Things are definitely beginning to stabilize a bit as we approach 0.6.
We're still planning to do some work on installers, and on various Quake 2
issues.  But before we start on that work, we want to nail down a few loose
ends.

This release improves support for testing code that uses (jump ...),
state-db.ss and data-file.ss.  It also clears the state-db on each reload,
and catches attempts to override an existing member function using ELEM.
These are all annoyances that we've encountered while stabilizing a large
Halyard-based project.

Eric Kidd (7):
      Clear gStateDB when script is reloaded
      Don't allow ELEM to clobber inherited methods
      Updater: Fix AUTO-UPDATE-POSSIBLE? to always use DIR
      Allow expressions to be passed to ASSERT-JUMPS
      Trap unexpected JUMPs while running unit tests
      state-db: Add better supporting for writing unit tests
      data-file: Add better support for writing unit tests