Quake 2: Don't write configuration or auto-saved games to disk
Under Vista, the directory "C:\Program Files" _appears_ to be writable, but
really isn't.  Instead, writing to this directory triggers a dodgy
compatibility layer.  We'd prefer to avoid this issue entirely, so this
patch removes all the disk I/O in Quake 2 that we know about.

We also remove the corresponding .gitignore entries.