Installer: Include On2_VP3.qtx again
Our default codec is the VP3 codec.  We need to include a copy of it in the
installer, and we used to, but the line was commented out.  This patch
uncomments that line so that we'll pick up the codec, and adds the
skipifsourcedoesntexist flag so we can still build installers if that
file is not present.

Reviewed-by: Robinson Tryon <robinson.p.tryon@dartmouth.edu>