0.5.25 - 5 Feb 2009
We're getting ready for an alpha release of one of our in-house programs,
so it's time to update the copyrights and license information, and fix
some bitrot in the installers and build system.

This release also modifies Quake 2 to stop writing automatic save files and
configuration data to disk.

Brian Campbell (5):
      Updated buildscript to latest version
      Updated buildscript to latest version
      Rakefile: updated 'sign' task to refer to current engine location
      Installer: Include On2_VP3.qtx again
      Merge branch 'master' of ssh://imlsrc.dartmouth.edu/var/lib/git/halyard

Eric Kidd (6):
      Quake 2: Don't write configuration or auto-saved games to disk
      Merge branch 'master' of ssh://imlsrc.dartmouth.edu/var/lib/git/halyard
      fix-prologues.pl: Update for 2009 and extend to *.rb files
      Update copyrights to 2009
      Don't use windows-installer.iss for all versions of *.iss file
      LICENSE.txt: Update for 2009 and latest changes