Rakefile: Call UpdateInstaller and buildscript tests during 'rake test'
In order to ensure that our unit tests don't bit rot quite as often, we
would like our unit tests to be run from our main 'rake test' task.  This
patch creates a new 'rake test_tools' task that runs the 'rake test' command
in the buildscript directory, and runs the update_installer_test.rb tests
directly (because the 'rake test' task would introduce a gratuitous extra
build of the UpdateInstaller.exe and UpdateInstallerTest.exe).

In order to support this, we need to move a few things around in the
UpdateInstaller Rakefile and tests.  In particular, we remove the extra
step of running the UpdateInstallerTest.exe by moving that into a test
within update_installer_test.rb.