Installer: Always install UpdateInstaller.exe in program root
In order for our old programs to be able to find and run a new version
of UpdateInstaller.exe that they download (which we do to ensure that we
can inject new code into the update process in case we need to do anything
tricky), we need to insure that our UpdateInstaller.exe is in the top level
of the program.  When downloading an update, that is where the updater
finds the update installer, and if it doesn't find it there it will just
use the old version that it already has.