buildfile.rb: Push engine binaries to Git repository
In order to support Git development for Halyard programs, it is
convenient to have a Git repository containing engine binaries, which
can be linked in to the project as a submodule.

In order to achieve this, without duplication of code, we can use the
'rake halyard:freeze' task for building an engine binary directory,
and then use that conistently for our Subversion binaries repo, our
Git binaries repo, and the binaries that we push to our halyard-test
zipfile.

In the process of developing this patch, I found it necessary to be
able to do dirty builds of the engine, for more rapid turnaround.
This patch also adds the necessary conditionals to allow dirty builds
to work; don't clone a new repo for dirty builds, delete directories
that are in the way of some commands, and change several mkdir
commands to mkdir_p.