buildfile.rb: Don't use git clone --no-checkout
'git clone --no-checkout', followed by a 'git checkout' of the build
version, had been used to avoid causing problems when there were file
renames that changed only case on a case-insensitive file system.  In
newer version of git, this does not work, because git sees the lack of
checkout as if every single file had been deleted, and refuses to do
the checkout.

This patch simply removes the --no-checkout; we will have to find
another fix or workaround the file rename problem if we hit that
again.