buildfile.rb: Remove Subversion support
We have officially migrated all of our projects off of Subversion.  This
patch removes all of the support for pushing our binaries to a Subversion
tag.  It also switches to our rake task for pulling gpgv.exe out of Git
instead of pulling it from Subversion, and removes all references to
Subversion in our comments.