HACKING.txt: remove outdated instructions, and mention modern layout
Our HACKING.txt had some old instructions about how to set up
buildscript, which is now just a submodule and the paths are
automatically set up.  Also, our "What Lives Where" section was fairly
out of date, so it has been updated to reflect the modern layout of
the project.