tools: Add announce-build.sh tool for emailing build announcements
When we first did our public release of Halyard, we tried to behave as
any public open source project with regular release announcements, a
Freshmeat project, etc.  As we got busy, however, we ceased updating
our release information with each build, since there was a lot of manual
labor involved.

This patch series adds a script, tools/announce-build.sh, which is
intended to automate the announcement of a new release of the engine.
It is separate from the build process, so that we can ensure a build
has worked successfully and can be used in a project before announcing
it publicly.

In this patch, the script composes an email, using the template we had
used for previous relase announcements, and attempts to put the draft
email in the user's IMAP drafts folder, using git imap-send (see help
for git imap-send for configuration details).