Crash reporter: Remove message about finding curl.exe
Some of the crash reports we're getting in from the field are coming
in with a message about where they found curl.exe as the report title,
as opposed to the actual erorr message.  This makes them sort poorly
in FogBugz, so this patch removes the debug log message that is
causing this.  We also remove a warning line when searching the
system path for curl.exe for good measure.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?14522