tools: Add irc announce bot, for announcing a build on IRC
Add a bot to announce a new version of the program on IRC.  The bot is
a reasonably generic bot that can be driven from standard input and
can identify to NickServ, join and part from channels, can send
messages to channels or users, and has a special purpose command for
performing a regexp search and replace on the topic of a channel, in
order to replace only the version number without changing the rest
of the topic.  See the documentation in tools/irc-announce.rb for
details on the command line options and commands supported.

This patch also hooks up the irc-announce bot to the rest of the
announce build infrastructure.  In order to allow testing of various
announcements without actually pushing out the announcements, this
patch adds the ability to specify only particular announcements to run
at a time, and fixes up the argument processing of announce-build.sh
and announce-build-freshmeat.rb to better be able to handle --verbose
and --dry-run options.  See the various usage messages for details.