Buildscript: Add tool for staging and releasing updates on server
This patch adds a new tool directory to the template directory hierarchy
that is populated by rake halyard:update_tools.  This tools directory
is intended to hold scripts used for managing a Halyard project.

This patch adds a new tool, release-updates, which is a thin wrapper
around the underlying functionality in Buildscript.  This is used for
managing releases of updates on the update server; it can ssh to the
update server, and copy spec files from various builds into the
release.spec and staging.spec for the updater, along with logging as
it goes to enhance our ability to audit our releases, as well as
giving us a history of what has happened in case we would like to
roll back.  It also ensures that each release is tagged based on the
build ID of the file being released, so that we won't forget to
tag updates that we've pushed out to tester.

Brian Campbell (11):
      Update server installer: Refactor tests to eliminate common code
      Update server installer: Make manifests and .spec files not writable
      Update server installer: Make pool files not writable
      Update server installer: Keep log of all changes to spec files
      Update server installer: Allow username to be passed in
      Update server installer: Remove unnecessary require
      Update server installer: Remove test case line that fails on Cygwin
      project_config.rb: Add new parser for project configuration
      project_config.rb: Add support for [sections], and better errors
      Add tools for staging and releasing updates on server
      commands.rb: Ignore .git directories instead of .svn when uploading update