eg: Centralizing (gensym) calls for anonymous elements
We want to make element names a bit more stable, so that Electric Gibbon
can keep track of which elements it tested on a previous visit to a
card.  That means getting rid of gensym'd element names, and replacing
them with more predictable names.  Unfortunately, (gensym) is used
widely throughout many libraries to generate element names, so before
replacing it with something else, we first need to centralize it.

We allow users to pass :name #f when creating an element.  This is a
little bit tricky, because the .name attr has type <symbol>, so we need
to do a little initialization magic.

PORTING: Please remove all gensym'd element names from existing code,
and replace them with #f.  Pay careful attention to new-FOO functions
that wrap .new.