eg: Split %test-planner% into %(null|shallow)-test-planner%
I've decided that we should keep the old buildbot behavior available for
people (1) who want a quick run-through of their program, or (2) who
haven't gotten the full run-through working yet.  The first step in
doing this is to create a %null-test-planner%.  This looks like the
regular test planner, but it doesn't actually run any test actions.

The next step: Hooking this up to the Rakefile and jump-to-each-card.ss.