eg: Reimplement events using Mizzen instead of Swindle
We've been meaning switch the events.ss module to Mizzen for a while.
And now that Electric Gibbon is about to start generating fake events
internally, it's time to stop procrastinating and actually do it.

Note that we clone the test suites in events.ss, and modify one copy to
test our newly-added backwards compatibility layer, and modify the other
to test the new official API.