eg: Refactor %test-planner% for clarity and fix bugs
After discussing %test-planner% and %deep-test-planner% today, Brian and
I decided that the current code was (a) far too hard to understand, and
(b) actually incorrect.

Our goals for this redesign:
 1) Centralize the running of test actions in %test-planner%.  This
    actually fixes two bugs:
    - We weren't adding "Rerun" actions to the actions-taken list.
    - We weren't draining the deferred thunk queue after a "Rerun"
      action.
 2) Remove the "iterator" interface in %test-planner%.  This caused an
    unnecessary separation between the code that did the planning and
    the code that ran the test actions, resulting in many long comments
    trying (unsuccessfully) to explain what was going on.

Note that we had to remove one test case from
%deep-test-planner-basic-test%.  Fortunately, the same functionality is
still coverted in %deep-test-planner-deep-test%.

Many thanks to Brian for pointing out the major issues with this piece
of code, and for discussing alternative designs.

Reviewed-by: Brian Campbell <brian.p.campbell@dartmouth.edu>