log: Combine gLog and gDebugLog into unified logging API
The problem: Our logging framework is very old, and it has accumulated
a large amount of "technical debt".  We want to add a number of new
features to the logging framework for use by Electric Gibbon.  We have
the choice of going further into technical debt, or of cleaning up this
section of code.  We've decided to go for the cleanup.

Our new API is based on log4j and log4cxx, two popular logging libraries.
Every log message has two values associated with it:

  level: One of trace, debug, info, warn, error and fatal.
  category: A period-separated string of module and submodule names.

Unlike our old API, the new API does not require our callers to know
which log file should actually contain a given message.  That decision
is now made automatically by the logging framework itself.

Major changes:
  - TLogger has been renamed to TLog.
  - The old gLog and gDebugLog are no longer public.  A new gLog object
    provides a unified interface to internal gHalyardLog and gDebugLog
    objects.
  - In Scheme, logging is now performed using either the low-level
    logger function or the high-level trace, debug, info, warn,
    log-error and fatal functions.  (We can't have an "error" function,
    because it already exists.)  Most of the old logging functions have
    been deprecated, and the rest will be deprecated shortly.
  - The EnvironmentError function is now handled using the
    "halyard.environment" category.
  - Logging call sites have been converted.  Some of these have been
    changed more than others, and some of the Debug callsites now use
    Trace.