log: Replace WITH-ERRORS-BLOCKED with WITH-EXCEPTIONS-BLOCKED
The old WITH-ERRORS-BLOCKED syntax would internally convert exceptions
into strings before passing them to the caller's REPORT-FN.  But this
required us to keep around the otherwise obsolete REPORT-ERROR and
FATAL-ERROR functions.

The new WITH-EXCEPTIONS-BLOCKED syntax passes the raw exception to its
caller's REPORT-FN.  Two new report functions are provided:
REPORT-EXCEPTION and REPORT-FATAL-EXCEPTION.  (In the future, I'd love
to add some support for log categories to these functions, but that will
probably require defining a new error class.)

The REPORT-ERROR and FATAL-ERROR functions are now deprecated, along
with the WITH-ERRORS-BLOCKED syntax.  Unit tests are provided for both
WITH-ERRORS-BLOCKED and WITH-EXCEPTIONS-BLOCKED.