log: Move static functions from TLog to TLogger
These functions are general support routines for displaying dialogs,
preparing to exit the program, and performing other similar tasks that
aren't related to a particular log file.  So we want to move these up
into TLogger.  This should allow us to move TLog out of the public
header in a future commit.