log: Move TLog into its own file
We want to get TLog out of the widely-used TLogger header, in
preparation for replacing it entirely.

Note that this patch changes the division of responsibility between
TLog::AlertBuffer and TLogger::DisplayAlert.  The latter is now checks
the value of TInterpreterManager::IsInCommandLineMode(), and decides
whether or not to use ConsoleAlert.

The change in behavior affects only one caller, the function
FancyCrashReporter::ReportCrashInCrashRepoter, which will now do the
right thing when the crash reporter crashes while running in command-
line mode.