log: Move vsnprintf (and buffer) into TLogger
We want to start moving logic out of TLog and into TLogger.  But to do
that, we need to construct a formatted log message in TLogger::vLog.  So
we move the log formatting code out of TLog, and modify all of TLog's
logging functions to take a std::string as an argument.

The new message formatting code uses a local, heap-allocated buffer to
avoid having log messages accidentally overwrite the buffer of _other_
log messages.  This tends to happen when making logging calls from the
FancyCrashReporter, either directly, or when unrelated Halyard
subsystems keep running during a crash.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?14522