log: Move warning inteception code to TLogger
Our unit tests need to be able to trap warnings and handle them
internally.  This has historically been handled by TLog::Warning.  This
patch moves this logic into TLogger::vLog, and generalizes it to support
a wider variety of log messages.

Note that the new code tries to be a bit more robust--it attempts to
block re-entrant calls to MaybeHandleLogMessage.