log: Remove old TLog class
Now that log4cplus is working, the TLog class is redundant.  By deleting
it, however, we do loose several features:

 1) Logs are no longer registered with the crash reporter.  This will be
    easy to fix.
 2) We always generate a Debug.log and Trace.log, even in runtime mode.
    Since we can now control the size of these logs, this isn't
    necessarily a problem.

With the removal of this code, we can _almost_ remove TDeveloperPrefs as
well.  The only remaining field in TDeveloperPrefs is the one which
controls whether the user can exit full screen mode.