Remove TDeveloperPrefs
Now that Debug.log is no longer under the control of TDeveloperPrefs, we
only use config/developer.prefs to control the ability to switch from
RUNTIME_MODE to AUTHORING_MODE on the fly.

This patch deletes TDeveloperPrefs, and adds a function
ENABLE-DEVELOPER-TOOLS-IN-ALL-MODES! which can be called by a script to
control whether or not users can switch from RUNTIME_MODE to
AUTHORING_MODE.

With this patch, the only *.cpp files left over in recognizable form
from pre-2002 5L are TPoint, TRect and TVariableManager.  Other
remaining files include the corresponding *.h files, and TCommon.h,
TPlatform.h and TVersion.h.

PORTING NOTES:

Remove config/developer.prefs and decide whether to use
ENABLE-DEVELOPER-TOOLS-IN-ALL-MODES!.