eg: Try to avoid infinite loops when replaying actions
Brian found an interesting bug while running Electric Gibbon, and
suggested this fix.  It turns out that Electric Gibbon would go into an
infinite loop if it tried to replay an action, and that action suddently
started calling JUMP when it hadn't before.

At his suggestion, I've added some minimal code to try and detect this
situation and abort the Electric Gibbon run.  We can probably think of a
better way to handle this, but it isn't a high priority right now.

Acked-by: Brian Campbell <brian.p.campbell@dartmouth.edu>