HACKING.txt: Clarify win32 build instructions
When trying to build Halyard in win32 I had an issue regarding the
order of the Libraries and Include files. This patch clarifies the
instructions to indicate that lines must be added to the end of the
lists, not the top as Visual Studio does by default.

I also clarified the test of another section to more accurately
describe how to navigate a dialog box.