TLogger::ConsoleAlert: Display "WARN:", not "WARNING:"
Right now, our buildbots are configured to look for "WARN:" in the
output of Halyard, because that's what out log4cplus configuration files
generate.  But on Windows, we never actually see log4cplus output on the
console, and instead rely on ConsoleAlert to display warnings.  (Yes,
this means that warnings will be printed twice on platforms with working
console output.)

This patch makes our ConsoleAlert output consistent with log4cplus, in
hopes of keeping buildbot configuration simple.