Add asynchronous %url-request% element
Our problem: We need to be able to perform JSON requests in the
background while running Halyard-based applications.

This patch is the first step towards that goal.  It provides a new
%url-request% element that implements asynchronous HTTP GET requests.
(Future versions will support POST, and most like other protocols such
as FTP.)

We implement asynchronous networking support using an (invisible)
element, because that gives us two major advantages:

 1) We get support for event-handling for free, by reusing the normal
    event-handling code for elements.
 2) %url-request% objects are automatically destroyed when their parent
    node is destroyed.

This code includes a very simple HTTP server for running unit tests.  To
run it, install the Ruby sinatra gem, and launch
tools/http-test-server.rb.  All the test cards depending on this server
live in /networking/tests instead of the regular /tests sequence, so that
we don't need to permanently keep this test server running in everybody's
development environment.

Most of the code for setting up CURL 'easy handles' in this patch is
adapted from Brian's work in wx/src/Downloader.cpp.

(This patch is fairly monolithic, because I started with some proof-of-
concept code, and cleaned it up until I liked it.  The resulting patch
series was a mess, so I squashed it down to a single commit.)