Installer: Pass "." as argument in shortcuts instead of full path
If a user moves their Halyard installation, Windows will update the shortcut,
but not the path to the Halyard project that we use to boot into runtime
mode.  This causes Halyard to crash with a mysterious error.

This patch changes the path used in the various shortcuts to be ".", which
works in recent version of Halyard 0.5.x.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?12997