proxy: Query for IE settings and log
This code is inspired by the answer to this question:

  http://stackoverflow.com/questions/202547/how-do-i-find-out-the-browsers-proxy-settings

Note that it doesn't do anything yet beyond extracting the configuring
information and dumping it to the log.

We now require winhttp.lib, which gives us access to the IE proxy
settings.  You will need to install Windows SDK v6.1 to get the
necessary headers and libraries.