TPrimitives: Move formatting of strings and booleans into TValue
Logging of primitive calls is currently handled by TArgumentList, which
queues each argument to be logged as it is read and logs them all (along
with the result) after the primitive has run.  This is rather baroque,
so we are trying to move the logging functionality out of TArgumentList
and into TPrimitiveManager::CallPrimitive().

This patch moves the formatting of string and boolean values from
TArgumentList into TValue, and adds a few unit tests to check that
it works.  In order for the unit tests to work, it exposes the
formerly private method ToDisplayValue().

v2: Use MakeQuotedString instead of appending quotes