TPrimitives: Remove TArgumentList::GetListArg() and corresponding operator>>
Among our >> operator definitons for TArgumentList, we had one that
would read in a list as a TArgumentList.  This was last used to
provide list primitives in old 5L, and has not been used since; all
other lists have been passed in as a variable number of arguments, or
as a TValueList stored in a TValue.  This patch removes those old
crusty vestiges of 5L.