TPrimitives: Suppres noisy primitive logging using log4cplus
We had been suppressing logging of certain frequently-called
primitives by calling SkipPrimitiveLogging while the primitives were
running.  We would like to start logging primitives when they are
called, rather than after they return, so we can tell which primitive
caused a crash or other bug.

In order to do this, we need to supress logging on the primitive
before we call it, so this patch removes SkipPrimitiveLogging and
instead moves the suppression to the log4cplus configuration, where
we set the loggers for all of the noisy primitives to a log level
of OFF.