TPrimitives: Log prim calls before and after they are executed
To get a better handle on certain crashes, we would like to log
primitive calls and their arguments before the primitive is executed,
as well as the return value after it completes.

This patch strips out the clunky old system that appended arguments to
a string as they were read out of the argument list in the prim call,
and instead simply implements the standard operator<< for printing a
TArgumentList to a std::ostream.