TPrimitives: Factor out many formulaic operator>> functions with a template
We a whole lot of completely formulaic operator>> definitions for our
TArgumentList class, along with the GetFooArg() methods they would
call.  This patch factors out all of the completely formulaic version
of this pattern into one templated operator>> definition.  This in
turn breaks the TDummyArgumentList that TStyleSheetTests.cpp used to
mock up an argument list, so instead we remove the dummy class
entirely and just builds up a normal TArgumentList by hand.