Build fix: libs/plt
After upgrading to the Windows SDK 6.1, PLT failed to build.  This patch
updates libs/plt with build fixes for the Windows SDK 6.1.  It also removes
the build of MrEd, which we have never used and which fails to build under
the new SDK.

Fix-for: 93b50e6e7c883ed5b4239e48371a20e227bc5845